requirements.txt
Flask
Flask_SQLAlchemy
PyJWT
bcrypt
